[Feat] 공통적으로 들어가는 메서드를 제공하는 BaseEventEditVC 적용
